# api-handler/openapi_travel_agency_sprint2_v1.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "OpenAPI Travel Agency (Sprint 2)",
    "version": "1.0"
  },
  "servers": [
      {
        "url": "http://localhost:8080/api/v1",
        "description": "local development"
      },
    {
        "url": "{protocol}://{host}:{port}/{basePath}",
        "description": "custom server",
        "variables": {
          "protocol": {
            "enum": ["http", "https"],
            "default": "https",
            "description": "The protocol to use"
          },
          "host": {
            "default": "api.example.com",
            "description": "The host name or IP address"
          },
          "port": {
            "enum": ["80", "443", "8080", "8443"],
            "default": "443",
            "description": "The port number"
          },
      "basePath": {
            "default": "api/v1",
            "description": "Api version"
          }
        }
      }
    ],
...
  "paths": {
...
    "/auth/forgot-password": {
      "post": {
        "tags": ["Auth"],
        "summary": "Password Recovery (User Story 13, Optional)",
        "description": "This endpoint sends a verification code to the user's email for password recovery.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  }
                },
                "required": ["email"]
              },
              "examples": {
                "example-1": {
                  "value": {
                    "email": "user@example.com"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Verification code sent to email.",
            "content": {
              "application/json": {
                "examples": {
                  "example-1": {
                    "value": {
                      "message": "Verification code sent to email."
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Email not found.",
            "content": {
              "application/json": {
                "examples": {
                  "example-1": {
                    "value": {
                      "error": "Email not found."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Failed to send verification code.",
            "content": {
              "application/json": {
                "examples": {
                  "example-1": {
                    "value": {
                      "error": "Failed to send verification code. Please try again later."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/verify-code": {
      "post": {
        "tags": ["Auth"],
        "summary": "Enter Verification Code (User Story 13, Optional)",
        "description": "This endpoint validates the verification code sent to the user's email.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "verificationCode": {
                    "type": "string"
                  }
                },
                "required": ["email", "verificationCode"]
              },
              "examples": {
                "example-1": {
                  "value": {
                    "email": "user@example.com",
                    "verificationCode": "123456"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Verification code validated successfully.",
            "content": {
              "application/json": {
                "examples": {
                  "example-1": {
                    "value": {
                      "message": "Verification code validated successfully."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid verification code.",
            "content": {
              "application/json": {
                "examples": {
                  "example-1": {
                    "value": {
                      "error": "Invalid verification code."
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Email not found.",
            "content": {
              "application/json": {
                "examples": {
                  "example-1": {
                    "value": {
                      "error": "Email not found."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Failed to validate verification code.",
            "content": {
              "application/json": {
                "examples": {
                  "example-1": {
                    "value": {
                      "error": "Failed to validate verification code. Please try again later."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/reset-password": {
      "post": {
        "tags": ["Auth"],
        "summary": "Enter New Password (User Story 13, Optional)",
        "description": "This endpoint allows the user to reset their password using the verification code.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "verificationCode": {
                    "type": "string"
                  },
                  "newPassword": {
                    "type": "string"
                  }
                },
                "required": ["email", "verificationCode", "newPassword"]
              },
              "examples": {
                "example-1": {
                  "value": {
                    "email": "user@example.com",
                    "verificationCode": "123456",
                    "newPassword": "NewSecurePassword123!"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset successfully.",
            "content": {
              "application/json": {
                "examples": {
                  "example-1": {
                    "value": {
                      "message": "Password reset successfully."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid verification code.",
            "content": {
              "application/json": {
                "examples": {
                  "example-1": {
                    "value": {
                      "error": "Invalid verification code."
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Email not found.",
            "content": {
              "application/json": {
                "examples": {
                  "example-1": {
                    "value": {
                      "error": "Email not found."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Failed to reset password.",
            "content": {
              "application/json": {
                "examples": {
                  "example-1": {
                    "value": {
                      "error": "Failed to reset password. Please try again later."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}